code update for inference_cogVideo_sat_refactor and arguments
